refactor: move specs types to types/xy_chart/utils
