+++ src/components/react_canvas/utils/rendering_props_utils.test.ts
@@ -1,4 +1,4 @@
-import { Rotation } from '../../../lib/series/specs';
+import { Rotation } from '../../../types/xy_chart/utils/specs';
 import {
   buildAreaRenderProps,
   buildBarRenderProps,
