+++ src/lib/series/domains/x_domain.test.ts
@@ -1,7 +1,7 @@
 import { getGroupId, getSpecId, SpecId } from '../../../utils/ids';
 import { ScaleType } from '../../../utils/scales/scales';
 import { getSplittedSeries } from '../series';
-import { BasicSeriesSpec } from '../specs';
+import { BasicSeriesSpec } from '../../../types/xy_chart/utils/specs';
 import { convertXScaleTypes, findMinInterval, mergeXDomain } from './x_domain';
 
 describe('X Domain', () => {
