+++ src/specs/settings.test.tsx
@@ -1,6 +1,6 @@
 import { mount } from 'enzyme';
 import * as React from 'react';
-import { Position, Rendering, Rotation } from '../lib/series/specs';
+import { Position, Rendering, Rotation } from '../types/xy_chart/utils/specs';
 import { DARK_THEME } from '../utils/themes/dark_theme';
 import { LIGHT_THEME } from '../utils/themes/light_theme';
 import { TooltipType } from '../types/xy_chart/utils/interactions';
