+++ src/state/test/interactions.test.ts
@@ -1,7 +1,7 @@
 import { BarGeometry } from '../../types/xy_chart/rendering/rendering';
 import { computeXScale, computeYScales } from '../../types/xy_chart/utils/scales';
 import { DataSeriesColorsValues } from '../../lib/series/series';
-import { BarSeriesSpec, BasicSeriesSpec, RectAnnotationSpec } from '../../lib/series/specs';
+import { BarSeriesSpec, BasicSeriesSpec, RectAnnotationSpec } from '../../types/xy_chart/utils/specs';
 import { getAnnotationId, getGroupId, getSpecId } from '../../utils/ids';
 import { TooltipType } from '../../types/xy_chart/utils/interactions';
 import { ScaleContinuous } from '../../utils/scales/scale_continuous';
