+++ src/types/xy_chart/specs/histogram_bar_series.tsx
@@ -1,6 +1,6 @@
 import { inject } from 'mobx-react';
 import { PureComponent } from 'react';
-import { HistogramBarSeriesSpec } from '../../../lib/series/specs';
+import { HistogramBarSeriesSpec } from '../utils/specs';
 import { getGroupId } from '../../../utils/ids';
 import { ScaleType } from '../../../utils/scales/scales';
 import { SpecProps } from '../../../specs/specs_parser';
